requirements.txt
playwright
jinja2
python-dotenv
Pillow
pytest
